1.x: make scan's delayed Producer independent of event serialization
It turns out serializing `request()` calls with regular `onXXX()` calls
can be problematic because a `request()` may trigger an emission of
events which then end up being queued (since `emitting == true`). If the
request is large and the queue otherwise unbounded, this will likely
cause OOME.

In case of `scan`, the fix was to make the missing request accounting
and arrival of the `Producer` independent of the event's emitter loop;
there is no need for them to be serialized in respect to each other.

In case of the `ProducerObserverArbiter` where the request accounting
and producer swapping has to be serialized with the value emission, the
solution is to call `request()` outside the emitter-loop.